GetStringAsyncEndpoints

+++ LearnMicrosoftHttpClient/HttpClientMethods/Dtos/GitHubAvatarResponseDto.cs
@@ -1,4 +1,4 @@
 ﻿namespace HttpClientMethods.Dtos
 {
-    public record GitHubAvatarResponseDto(string Url, string ContentType, long ContentLength);
+    public record GitHubAvatarResponseDto(string Url, string ContentType, int ContentLength);
 }
